// front-end/src/app/Telas/admin/models/vagasModel.ts
export class VagasAdmin {
    id: string;
    adminId: string;
    funcao: string;
    descricao: string;
    area: string;
    horario: string;
    salario: number;
    beneficios: string;
    quantidade: number;
}

export class DadosAdmin {
    id: string;
    nome: string;
    email: string;
    empresa: string;
    cargo: string;
}